requirements.txt.txt
python-telegram-bot
ccxt
requests